main/boost: fix and reenable failing debugger tests In debugger::debug_parent_wait() the variable ch was mistakenly defined to be a char and used to obtain a value from ch = fgetc( command_child );
Later in the code a compare of the truncated ch (0xff) to the constant EOF (0xffffffff) would fail resulting in errors between the expected and actual values returned in the debugger.py and debugger-mi.py tests.

Since fgetc is defined to return an int the declaration of ch was modified and the truncation eliminated.

I will post the patch to the boost/build community as well.