main/mariadb: upgrade to 10.3.9 and move default config
move default config to /etc/my.cnf and /etc/my.cnf.d/. This is to make
it easier to ship default configs for clients, plugins and mysql,
without introducing conflicts.

We also add a post-upgrade script which makes sure that the
skip-networking setting does not disable networking in a pre-existing
config on upgrade.