main/busybox: use the non-standard timegm(3) for sysklogd -Z
While it is not standardized by POSIX and a GNU extension musl has
support for it since 0.9.2 so we might as well use it.

See: https://bugs.alpinelinux.org/issues/9279#note-5

Fixes #9279
Fixes #7818

While at it also make sure that tm_year is initialized correctly (the
problem here being that the year is not part of the strptime format
string).

Fixes #8028