testing/singularity: remove maintainer
The 3.0 release is a major rewrite, which involves some bizarre new
build process, additional dependencies, a license change, and removes
most of the features in exchange for other, different features.  In
addition, this is now a hybrid-license project trying to sell a "Pro"
upgrade.  I'm not going to maintain this going forward.