testing/i2pd: Allow for use with syslog
Currently the init script for i2pd expects I2PD_LOG to be set and
proceeds to touch the file on service start. However, i2pd is allowed to
log to stdout/syslog which does not require this log file to be created.
This patch removes the mandation that I2PD_LOG be set in
/etc/conf.d/i2pd and will only touch the file if the log variable is
set. For anyone using the default log-to-file settings, there is no
noticeable change in functionality.