main/acf-core: update lua-json4 depends
commit 66573786035d (main/lua-json4: install into lua/common, remove
subpkgs, add rock_manifest) renamed the lua5.x-json4 packages and added
a provides for backwards compatibility. But the build time dependencies
does not resolve the provides so builder will fail figure out the build
time order.

We work aorund this by adding an explicit depend of lua-json4, without
the lua version, so we no longer depend on proper provides resolving.