community/upx: upgrade to 3.95
UPX 3.94 seems completely broken even on x86_64 due to toolchain
upgrade.

 - use cleaner method to force CXXFLAGS precedence over "-O2";
 - add basic check() -- passes on x86_64 and mipseln8hf, may need
   tweak for other architectures;

Just for the record: out-of-tree hard-float mips* is still broken
for any application that uses FPU because UPX does not copy the
attribute section to the compressed binary so the usual "FR=1 code
in FR=0 context" problem arises. In-tree soft-float mips* should
not be affected.