community/go: enable build on armel
There were ARMv5TE boards with 512 MiB (as in RPi1), that is enough
for go.

Also remove go-bootstrap from makedepends. In a cross-build it ends
up with depending on itself which is safe (abuild handles this) but
ugly. In normal build it is duplicated which is safe but ugly too.

Do not bump pkgrel since
 - no previous armel packages exist;
 - makedepends change makes no difference for the created apks.