testing/nginx-mod-http-modsecurity: new aport
This is the nginx connector for ModSecurity. It depends on the newly
added aport testing/modsecurity. I dropped this into testing, as it
is new and would depend on a package outside of main. However, it
really should become part of the main/nginx APKBUILD eventually.
Until it does, the APKBUILD here will need to be bumped when nginx is
in order for the module to load.