community/snapcast: fix OpenRC service
Without creating /var/lib/snapserver the services fails to start with
the following error message:

	Exception: failed to create settings directory: "/var/lib/snapserver/": 13

Because the snapcast process itself tries to create the directory but
doesn't run as root and therefore gets an EACCESS (13) error.

While at it I also specified the pidfile created by the snapcast
process. No idea how OpenRC was even able to track the service status
without it.