main/syslog-ng: replace weirdly modularized configs with single file
Building syslog-ng configuration from multiple fragment files using
a shell script is unnecessary complicated and error prone - syslog-ng
supports the include directive.

This change is not backward compatible, but I quite doubt that anyone
was using this.

The new config is based on the default config for RedHat from upstream.