main/mkinitfs: fix zfs root with grub
grub-mkconfig will generate a root=ZFS=pool/path line. Fix nlplug-findfs
to deal with search device ZFS=... and fix initramfs to also deal with a
missing rootfstype=zfs

A side-effect of this is that it is faster to boot zfs root since we no
longer need to wait for nlplug-findfs timeout of 5 seconds.