main/libtls-standalone: Remove dependency to fix bootstrap
The package depends on ca-certificates-cacert, which breaks
bootstraping a cross compiler toolchain:
busybox -> libtls-standalone -> ca-certificates-cacert

The last package haven't been built so far, so the bootstrap
process aborts. The ca-certificates package cannot simply
be built by the bootstrap script due to a wrong import statement
for the c_rehash program (which, btw, we don't need at this point).

Why does the libtls-standalone have a dependency to the ca-certificates in first place?