main/dhcp: correct license
The license was changed when 4.4.0 was released.

Fixes https://bugs.alpinelinux.org/issues/10161

DHCP is available for free download under the terms of the MPL 2.0 license.
https://www.mozilla.org/en-US/MPL/2.0/

https://www.isc.org/downloads/software-support-policy/isc-license/
"BIND versions prior to BIND 9.11.0, all versions of ISC DHCP prior to ISC DHCP 4.4.0, and Kea versions prior to Kea 1.0.0 are released under the ISC License[...]"