testing/redo: fix test relying on tty
One test fails on the builders because it relies on a tty being present.
Upstream has fixed this test to not rely on a tty anymore, but this has
not been released yet.

Apply the commit from upstream as a patch until a new version has been
released.

The previous patch to fix it does not work, so remove it.