main/openrc: add zsh and bash completion subpackages
The MKZSHCOMP and MKBASHCOMP variables need to be set for both make and
make install thus I simply exported them as environment variables in
build(). The alternative would have been passing them locally twice
which didn't seem very appealing to me. Not sure if we have a clear
policy on this though...