testing/frr: Update to 7.0
Insert myself as a maintainer. I have not heard from Arthur Jones for
quite some time. Given that there has been no update to the package
since August 2018 although upstream released two new major versions in
the meantime, this package seems abandoned to me, so I hope this is
acceptable.

FRR should build and run fine on all architectures, so the package
`arch` should not be limited to x86_64.

While a small part of FRR's source is provided under the terms of the
LGPL license, the resulting binary can only be distributed under the
terms of the GPL, version 2.0 or (at the users option) any later
version. Update the `license` field to reflect this.

Starting from version 6.0, FRR requires libyang, so add it as a
build time dependency.

FRR has modules for integrating SNMP and RPKI. Add their dependencies
as build time dependencies and package them as subpackages.

Add `noarch` to those subpackages which do not contain any architecture
specific files.

FRR changed its statup script to  `/usr/lib/frr/frrinit.sh` - propagate this
change to the OpenRC wrapper script.

`inet_ntop` in musl has been fixed to be standards compliant, so the
patch `make-check-bypass-isisd-tlv.patch` is no longer needed.

According to upstream the headers and shared libraries are not for
consumption by external programs, so they should not be packaged
as a `-dev` package.