main/git: Add disabled check option
We are currently not using checks on the build, as they fail anyway on
t/t3900-i18n-commit.sh on the Japanese fonts. So clearly state this in
the !check as per aports guidelines.

Also, if the tests would not fail, we probably better run a profiled
build instead which does testing explicitly.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>