community/mtd-utils: make find invocations compliant
The Busybox version of find allows parameters to be specified between
'{}' and '+' as argument to -exec. This however is not according to the
posix specification, and the GNU version of find doesn't accept it.

Use the coreutils version of 'mv' with the '-t' arguement to be able to
specify the filenames as last.