main/rng-tools: move from community
Maintaining the system entropy pool with a HWRNG is quite important on embedded
devices (e.g. RPI3) and on virtual instances, due to that these systems
generate very little entropy on their own.

Starting rng-tools early in the boot of a system with a hwrng avoids that the
Linux kernel "blocks" for a couple of minutes during boot, before any services
using entropy can start.

This package supports feeding the kernel entopy pool with random data from
different sources, among these /dev/hwrng and the Intel rdrand instruction.

The reason for moving it to main is that rng-tools for these reasons
also will be beneficial to provide on the Alpine installer media.