testing/procs: new aport
A modern replacement for ps written in Rust

re: tests
In the following scenarios, tests pass:
- running by hand on real hardware
- running by hand in an alpine container

However, the moment abuild is involved, tests will segfault on test 3.
I couldn't figure out why exactly this happens.
I have a hunch that it has something to do with stdout handling, but
it's hard to verify.