community/rust: build on arm{hf,v7}, ppc64le & x86
Also disable full bootstrapping. Since we always bootstrap from $pkgver minus one
we don't need it and it increases buildtime considerably.

Also keep in mind that we have to add new arches in multiple steps:
    1. Compile with the upstream triplets, compiling alpine's triplets in
    2. Compile again, now with our triplets selected as build/target, now that
       rustc knows about them

This means that the produced compilers MUST ONLY be used to bootstrap the actual
compilers with our triplets, DO NOT use these to compile packages! Since x86_64
still uses the usual triplets it's fine to upgrade/add packages for that arch
though, please DO NOT do this for armhf,armv7,aarch64,ppc64le or x86 though!