main/postgresql: create postgres account
This change was made due to complains that /etc/passwd contains system
accounts with login shells, postgres being one of them [0].

Setting the shell of the postgres account to /sbin/nologin was not
deemed feasible as the official document suggests login in as the
postgres user via su(1) [1]. Other Linux distribution (Debian, Gentoo,
Arch) also create the account with a shell.

Not creating the postgres account by default (via alpine-baselayout) and
instead creating it on demand from the package was deemed a reasonable
compromise.

[0]: https://gitlab.alpinelinux.org/alpine/aports/issues/10579
[1]: https://wiki.postgresql.org/wiki/First_steps