main/sdl2: fix directfb cflags in .pc
make sure the directfb CFLAGS are included in the .pc. This fixes the
following compiler error:

/usr/include/SDL2/SDL_syswm.h:77:10: fatal error: directfb.h: No such
file or directory
 #include <directfb.h>
           ^~~~~~~~~~~~

We don't add the directfb pkg-config file as a dependency because
directfb will not be directly linked to sdl, instead sdl will dlopen it.
We do need the headers though, so we simply add the directfb CFLAGS.

This is also what the autotools configure script does.