main/dns-root-hints: fail on changed signature too
In #alpine-devel, the possibility of an updated signed file with
the same serial was raised. To prevent this from giving a build
with potentially changed data a check on the signature data was
added.

I fully expect that this never occurs, but the check costs very
little and we are better off safe than sorry.