testing/ldc: new aport
This add a package to support the D Programming Language.
Currently, one has to jump through hoops in order to use D on Alpine.
This packages installs the official (and only) LDC build that came
with an Alpine build, LDC-1.8.0.
While it is an old version of the compiler, 2.078.3,
it can be used to bootstrap other compiler versions.