community/kodi: allow running kodi-gbm at boot as it's own user
If kodi-gbm gets installed, it creates the "kodi" system user.
Then using the service file Kodi can be started on boot to run in KMS
mode (kodi-gbm) as the kodi user. In combination with a polkit rule in
kodi-polkit the user can then shut down the system from within Kodi.

With this setup Alpine Linux with Kodi becomes viable for HTPC devices
and smart TV's.