main/linux-vanilla: avoid pulling in all firmware during build
We need to have linux-firmware in makedepends because the runtime
dependency is addedin package() function, and we need to make sure that
build order resolver gets that it needs to build linux-firmware before
the linux-vanilla.

Make it possible for build servers and lxc containers etc to manually
install linux-firmware-none to avoid pull in all linux firmware
dependencies during build.