testing/cachefilesd-inotify: new aport
This is a fork of community/cachefilesd that has been patched to work with INOTIFY instead of DNOTIFY (which is disabled in most Alpine Linux kernels).

It is hoped that this fork will eventually be brought back into cachefilesd's mainline codebase, but in the meantime, this package should unblock those who might have wanted to use cachefilesd with Alpine Linux.