main/bacula: use root as rundir owner for bacula-client
Partially revert previous commit

In case other bacula daemons run on the same host,
their package will create the bacula user and group
and their init script will set bacula as the owner of the rundir.
bacula-fd will be able to write the pidfile in the same rundir since it runs as root.