gitlab-ci: remove git fetch before_script
To know what aports are changed in a merge request, we need to know
where the target branch is, so that we can use git diff against that
branch.

To test this, this was added in the CI jobs description in a global
`before_script`, but this is now being handled by respective scripts
themselves, so the before_script is now redundant.