testing/vpnc: remove aport
upstream is dead, last release is from 2008. Additionally, the script
doesn't seem to work properly. When using it with testing/openconnect
I get the following error

	Error: either "to" is duplicate, or "uid" is a garbage.

This is due to the fact that /etc/vpnc/vpnc-script calls `ip route add`
with `uid 0` which is incorrect. While this could be fixed, removing the
aport makes more sense to me as upstream is dead. Since nobody seems to
have reported this issue I am assuming that this aport isn't used
anyhow.

Alternative solution: Package a fork. For example:

	* http://git.infradead.org/users/dwmw2/vpnc-scripts.git
	* https://github.com/streambinder/vpnc
	* …