main/gcc: Add support for the D programming language
GDC, the GCC-based D compiler, is part of GCC since GCC9.
Upstream support for Musl was added progressively (D relies heavily on libc),
and some key components were missing as of GCC 9.2.0 release (2019-08-12).
The final patches to support Musl have been  submitted to
the gcc-patches mailing list on the 2019-08-21.
This includes a smaller version of those patches,
tailored to only include the changes necessary to Musl.
Most of the patch could be removed after/if GCC 9.3.0 is released.