community/lua-cliargs: change pkgver scheme
Version number of Lua rocks consists of version of the package followed by
revision of the rockspec, e.g. 3.0-1. Authors of Lua packages usually don't
consider the revision number as part of the package version, but not in this
case. I've originally decided to represent the rockspec revision as an
additional digit in the pkgver (3.0-1 -> 3.0.1), but it has showed up that
it's causing confusions. I hope that 3.0_p1 will be less confusing.

I'm aware of that 3.0.1 > 3.0_p1 but the new version 3.0-2 didn't introduce
any notable changes and I don't expect 3.0-3 anytime soon (the project is
inactive since April 2018), so it shouldn't cause any problems.