testing/ldc: Install ldc-static alongside compiler
The previous dependency means that 'ldc' would not work out of the box.
One had to install ldc and ldc-static, or ldc and ldc-runtime and compile with flags.
This makes sure that ldc-static is installed so basic compilation works out of the box.
People wanting to compile with shared libraries will still need
to install the 'runtime' package explicitly.