main/bind: depend on bind-tools
The install script needs to call rndc-confgen to generate a key which is
required for reloading to work.

955e379151 (main/bind: add support for "reload", 2019-10-19) added
support for reload, but that causes an issues when installing bind,
because rndc-confgen is in the bind-tools subpackage, which is not
available if you just install bind.

Fix this by having bind depend on bind-tools.

Fixes #10935