community/py3-automat: re-enable tests but don't run ones for Twisted
Twisted is used for a subsection of functionality and cause a circular
dependency since Twisted uses automat as well, Graphviz is also removed
since it is the other dependency required for the same subsect of
functionality so no reason to keep it if we are going to disable testing
it.

This was found during the Python3.8 rebuild.