testing/openv2g: New aport
OpenV2G is an open source implementation of the ISO IEC 15118 and also
the DIN 70121 vehicle to grid (V2G) communication interface.

The main component are the libraries, but the repository also includes a
testing application.

The makefile(s) are a bit of a mess, and should be rewritten. However as
that is outside of the scope, only the most minimal changes where done
to produce a useable build.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>