main/ndctl: fix build version
Due to lack of a version file, ndctl tries to get a version from the
current git repository, which is aports. This leads to build failures,
because the version that it ends up with is invalid:

```
ERROR: ndctl-dev*: usr/lib/pkgconfig/libdaxctl.pc: pkgconf version
67.git410c029f is invalid
```

This can be fixed by creating a `version` file with the actual version
in it.