main/dbus: Remove uuid generation from post-install
In the past, dbus uuid generation was done as a post install step. Years
later, this was copied/moved into the start-up script. This makes a lot
of sense, where in the post-install step it does not.

More importantly however, when bootstrapping alpine into a directory,
with the dbus package, dbus-uuidgen fails during the post-install phase
due to missing /dev/urandom, which is expected not to exist in a
alternative root, as can be seen in the following error

  Failed to generate UUID: Could not open /dev/urandom: No such file or directory
  ERROR: dbus-1.12.16-r0.post-install: script exited with error 1

Lets drop this line all together as it serves no purpose with the UUID
being generated at startup anyway.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>