main/bison: upgrade to 3.5
The parallel build fix patch does no longer apply cleanly. Instead of
patching the build system, which seems to be an error-prone and
laborious process to me, simply disable parallel builds for this aport.
Also the bottleneck (in terms of time it takes to build the aport) isn't
the build() it's the check() function.