community/py3-tornado5: new aport
Salt does not support Tornado 6 yet, so we need to package Tornado 5.

This package is adjusted so that it can be installed alongside
py3-tornado. This is done by renaming the module for tornado to
tornado5. The test suite passes, except for some error that also happen
without renaming the module, and the test suite has been disabled
anyway.