main/gcc: Fix D language support on non-x86_64
While some basic functions were working, some C definitions in Druntime
were wrong, which led to wrong compilation when trying to bootstrap
other packages on x86, ARM and AArch64.

Additionally, s390x didn't seem to pick the dependency on the C runtime,
and needed an extra dependency on libucontext-dev
(not to mention a few fixes to C definitions, like other archs).

Finally, 404-dlang-zlib.patch contains a fix that was done in 2.077.0,
but is needed for another Alpine package to work on all platforms.
While not required to fix platform support, it fixes #11125.