community/mpd: run checkpath with quiet flag
The checkpath invocation creates a directory usually located on a tmpfs
and cleared on each boot. As such, mpd creates and "corrects" the
permissions of this directory on each boot. This causes two unnecessary
messages to be emitted on each boot by the service. This change adds the
quiet flag which suppresses these messages.