community/py3-google-auth: upgrade to 1.11.0
Dependencies are listed at:
https://github.com/googleapis/google-auth-library-python/blob/release-v1.11.0/setup.py#L21

Test dependencies are listed at
https://github.com/googleapis/google-auth-library-python/blob/master/noxfile.py#L17

This commit update the dependencies to match the above. The exception
continues to be the apparently optional py3-grpcio dependency which is
still in testing/.

Mock is a backport of unittest.mock to earlier Python versions. It isn't
required on Python 3.8, so this commit patches the tests to use the
standard library.