community/py3-ruamel.yaml: upgrade to 0.16.6
The previous homepage responds with 404. The upstream project, including
a mercurial repository, has moved from BitBucket to Sourceforge.

The [Sourceforge page] listed several files to download including:

    https://downloads.sourceforge.net/ruamel-yaml/ruamel.yaml-0.16.6.tar.xz
    fc305499241de6cf5c2fa694a53762d853dce44af34f4aff1774d0ee772620cea7398591d1f8f4c43573d276ffd8cfd147dae0e98764e1a5a6045542514dcc42

Which was enough to make sure check() succeeds:

The issue was that python failed to import ruamel.yaml: the solution is
to call `pytest-3` rather than `python3 -m pytest`

According to the [pytest documentation], a difference between running
pytest as `pytest` vs `python3 -m pytest` is that the later will also
add the current directory, in this case `$builddir` to sys.path.

Somehow if `$builddir` is on `sys.path` the package to test is not
found.

The Sourceforge download URL above now 404s and the project's files page
is now empty. PyPI can be used as `$source` instead. The artifact on
PyPI does not include the tests, so these are again disabled.

The C module is now distributed separately.

[Sourceforge page]: https://sourceforge.net/projects/ruamel-yaml/files/
[pytest documentation]: https://docs.pytest.org/en/latest/usage.html#calling-pytest-through-python-m-pytest