community/py3-ruamel.yaml.clib: new aport
Upstream split the C code out of community/py3-ruamel.yaml into a
separate package. In order for py3-ruamel.yaml and
ansible-lint to stay in community this separate package needs to be in
community.

There are no tests enabled because they are not available in a stable
download. The PyPI download doesn't include the tests. Although the
[project activity page] on Sourceforge shows files released, the [files
page] is empty.

[project activity page]: https://sourceforge.net/projects/ruamel-yaml-clib/
[files page]: https://sourceforge.net/projects/ruamel-yaml-clib/files/