community/mpd: enable avahi support
adding avahi-dev dependency and relevant compile option.
test/run_avahi fails to build, and could not find how to fix it. As this is just for build testing I remove it with remove-avahi-test.patch  Equally happy if someone finds a way to fix it...

Tested on x86_64 target system and mpd service is well advertised by avahi