testing/beancount: disable warnings and a test
This change disables warnings from pytest. The warnings results in
verbose output from `check()` including a list of test names in the
logs. After disabling the warnings, the last 500 KiB of the logs
as shown in the GitLab web UI becomes useful again and will include the
details of failing tests.

This change also disables a test that fails intermittently.