community/upx: disable on s390x, skip -static test on x86
the binary is completely useless on s390x so remove it rather than just
ignore the tests.

the x86 -static test(s) fails, but the dyniamic still works so lets only
disable the test for now to unblock the builders.

also let the test show error message while at it