main/openrc: don't shadow /lib/modules if modloop is an overlay
In the run-from-RAM scenario, the modloop init script runs after
APK has already installed packages. If any of these packages
contain files that live under /lib/modules, /etc/init.d/modloop
will hide them by mounting the modloop on top of this directory,
even if overlayfs is available.

This change places the pre-modloop contents of /lib/modules into
the overlay stack, on top of the actual contents of the modloop
file, so that in a conflict package contents override modloop
contents.