main/ncurses: fix automatic dependency due to symlinks
The -libs subpackage depended on ncurses-terminfo when it shouldn't
because when moving $pkgdir/usr/lib also moved /usr/lib/terminfo which
is a symlink to /usr/share/terminfo which is part of ncurses-terminfo
and thus a dependency would be automatically added by the
prepare_symlink() function from APKBUILD.

So make the libs() function grab only actual libraries by matching
/usr/lib/*.so.* (this does not match libraries meant for -dev which
would be *.so) so /usr/lib/terminfo isn't caught up.

Also make the terminfo() function move /usr/lib/terminfo to it.

In the meantime also do other cleanup like removing unnecessary 'cd
"$builddir"' declarations and make libs() depend on the same
$pkgver-r$pkgrel of the ncurses-terminfo-base.