main/ncurses: re-arrange terminfo contents
This re-arranges the contents of ncurses-terminfo-base so it contains
all the terminfo entries for commonly used TERMs like xterm256-color.

It also includes all terminfo entries for terminals we package like
alacritty, gnome-terminal, konsole, kitty, st, etc. And a few others
like putty, tmux and vte.

The ncurses-terminfo-base packaged increased from 94.21 kB to 126.98 kB,
a quite small increase when considering that now it is in the majority
of cases the only package one has to install instead of having to rely
on the massive ncurses-terminfo package.