community/py3-pyclipper: use clipper library from Alpine package
To save installation space, and to use the exact same implementation
of polygon clipping across the system.

Also taking the source code from pypi instead of github, like most
other Alpine py3 packages. Other than github, pypi is putting the git
release tag into the zip bundles, which is then used by Python
setuptools. When setuptools is installing a package, it is checking
for the version tag, and the installation fails with a warning when
the version tag cannot be inferred.  Before this change, the `APKBUILD`
would run `git init` to work around this setuptools check.  After this
change, that hack is not needed anymore.