main/openvpn: Add instance name to OpenVPN --daemon param
The --daemon OpenVPN params also controls the tag OpenVPN uses to ship
logs to syslog. If multiple OpenVPN instances are used, the user may
want to apply syslog filters, to have different log files for the
different instances.
If you specify an additional --daemon option in the config file, it will
be ignored, since the one on the command line is preferred.