community/rxvt-unicode: Remove hardcoded TERMINFO
If `--with-terminfo` is passed, urxvt sets the `TERMINFO` environment
variable to the passed value. The value is prepended to the terminfo
directory search order, i.e. it takes precedence over other directories
in the default ncurses search order.  For this reason, with
`--with-terminfo` enabled, `~/.terminfo` can no longer be used to
override terminfo files.

This change removes the `--with-terminfo` configure flag, thereby
ensuring that urxvt uses the default ncurses search order instead.

See also: #11304.