main/linux-lts: upgrade to 5.4.34 and misc config fixes
enable FS_VERITY_BUILTIN_SIGNATURES on all flavors

enable NET_REDIRECT for all architectures for which is default and
supported

add Southern Islands (SI) and Sea Islands (CIK) support for amdgpu
driver on x86 and x86_64

enable PKCS8 for architectures for which had it enabled.

enable other default crypto options on architectures which had it
disabled.