main/meson: add opinionated helper for Alpine Linux packaging
This creates an alpine-meson binary that is a shellscript that carries
the default values we think are best for packaging stuff in Alpine
Linux, they can of course be overridden in the APKBUILD by passing the
same options at the end.

It uses the same concept as the arch-meson from Arch Linux, the extra
values are taken from https://mesonbuild.com/Builtin-options.html and
from some peeking at Void Linux's meson build-style.

This should allow us to make sweeping changes to APKBUILDs with a single
change in a knob instead of on every template, such as the recent
changes to --buildtype=plain and CMake's -DCMAKE_BUILD_TYPE=None.

auto-features should be disabled but that would lead some packages like
gstreamer requiring the APKBUILD writer to switch up more than 60 lines.