community/py3-werkzeug: upgrade to 1.0.1
Also enable tests.

One test in tests/test_debug.py depends on py3-pytest-timeout which is
not available in the community repo. The timeout mark is applied as a
decorator which means that the error appears during collection and the
individual test can't be skipped. The check function ignores all tests
in that file.