testing/openv2g: Make the open-v2g library more versatile
The Open-V2G library can be made a little more versatile by enabling
more feature, such as XMLDSIG support.

More importantly, the library currently violates the ISO15118 spec by
using a home-grown UCS implementation, where everything is treated as
32-bits, e.g. UTF-32, where ISO15118 uses UTF-8 encoding. As this is
incompatible, we are better off switching to ASCII for now, a that
at least is backwards compatible.

To properly fix this, the library needs toe be fixed to properly by a
proper UCS implementation or using a library to this end.

Note, that these patches are configuration options changing the
defaults, and thus are not up streamed.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>